chore: move component
