+++ src/new-components/leadbox/leadbox.stories.mdx
@@ -1,5 +1,5 @@
-import { Leadbox, LeadLink } from './leadbox';
-import { ArgsTable, Canvas, Meta, Story } from '@storybook/addon-docs'; import { IllustrationSet } from '../illustration/illustration';
+import { Leadbox } from './leadbox';
+import { ArgsTable, Canvas, Meta, Story } from '@storybook/addon-docs';
 
 <Meta
   title="Components/Leadbox"
@@ -15,7 +15,7 @@ Give the user for further actions. This can either be a contact to approach or a
       illustration={'space_shuttle_043'}
       title="Haben Sie Interesse bekommen?"
       contact={{
-        email: "gholam.abdol@satellytes.com",
+        email: "example@satellytes.com",
         title: "Gholam Abdol, Geschäftsführer",
         headline: "Projektanfrage"
       }}/>
