+++ src/templates/blog-post.tsx
@@ -16,7 +16,7 @@ import { LocalesQuery } from '../pages';
 import FollowPanel from '../components/social-panel/follow-panel';
 import { useTranslation } from 'gatsby-plugin-react-i18next';
 import { Astronaut } from '../components/icons/illustrations/astronaut';
-import { LeadboxProps } from '../components/leadbox/leadbox';
+import { LeadboxProps } from '../new-components/leadbox/leadbox';
 
 interface BlogArticleTemplateProps {
   data: {
